+++ ingestion/src/metadata/parsers/schema_parsers.py
@@ -15,9 +15,6 @@ Hosts the singledispatch to get the schema parsers
 from typing import List, Optional
 
 from metadata.generated.schema.type.schema import FieldModel, SchemaType
-from metadata.parsers.avro_parser import parse_avro_schema
-from metadata.parsers.json_schema_parser import parse_json_schema
-from metadata.parsers.protobuf_parser import ProtobufParser, ProtobufParserConfig
 from metadata.utils.dispatch import enum_register
 
 schema_parser_config_registry = enum_register()
@@ -29,10 +26,14 @@ class InvalidSchemaTypeException(Exception):
     """
 
 
+# Load parsers only on demand
+# pylint: disable=import-outside-toplevel
 @schema_parser_config_registry.add(SchemaType.Avro.value.lower())
 def load_avro_parser(
     topic_name: str, schema_text: str  # pylint: disable=unused-argument
 ) -> Optional[List[FieldModel]]:
+    from metadata.parsers.avro_parser import parse_avro_schema
+
     return parse_avro_schema(schema_text)
 
 
@@ -40,6 +41,8 @@ def load_avro_parser(
 def load_protobuf_parser(
     topic_name: str, schema_text: str
 ) -> Optional[List[FieldModel]]:
+    from metadata.parsers.protobuf_parser import ProtobufParser, ProtobufParserConfig
+
     protobuf_parser = ProtobufParser(
         config=ProtobufParserConfig(schema_name=topic_name, schema_text=schema_text)
     )
@@ -50,6 +53,8 @@ def load_protobuf_parser(
 def load_json_schema_parser(
     topic_name: str, schema_text: str  # pylint: disable=unused-argument
 ) -> Optional[List[FieldModel]]:
+    from metadata.parsers.json_schema_parser import parse_json_schema
+
     return parse_json_schema(schema_text)
